Completes autoconfigure module
Fixes #262
