+++ spring-cloud-gcp-autoconfigure/src/main/resources/META-INF/spring.factories
@@ -1,5 +1,10 @@
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.cloud.gcp.autoconfigure.core.GcpContextAutoConfiguration
+org.springframework.cloud.gcp.autoconfigure.core.GcpContextAutoConfiguration,\
+org.springframework.cloud.gcp.autoconfigure.logging.StackdriverLoggingAutoConfiguration,\
+org.springframework.cloud.gcp.autoconfigure.pubsub.GcpPubSubAutoConfiguration,\
+org.springframework.cloud.gcp.autoconfigure.sql.GcpCloudSqlAutoConfiguration,\
+org.springframework.cloud.gcp.autoconfigure.storage.GcpStorageAutoConfiguration,\
+org.springframework.cloud.gcp.autoconfigure.trace.StackdriverTraceAutoConfiguration
 
 org.springframework.cloud.bootstrap.BootstrapConfiguration=\
 org.springframework.cloud.gcp.autoconfigure.config.GcpConfigAutoConfiguration,\
