+++ spring-cloud-gcp-starters/spring-cloud-gcp-starter-logging/pom.xml
@@ -21,10 +21,5 @@
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-gcp-logging</artifactId>
 		</dependency>
-		<dependency>
-			<groupId>javax.servlet</groupId>
-			<artifactId>javax.servlet-api</artifactId>
-			<scope>test</scope>
-		</dependency>
 	</dependencies>
 </project>
