+++ spring-cloud-gcp-starters/spring-cloud-gcp-starter-pubsub/pom.xml
@@ -25,11 +25,5 @@
 			<groupId>org.springframework.cloud</groupId>
 			<artifactId>spring-cloud-gcp-pubsub</artifactId>
 		</dependency>
-		<!-- @ConfigurationProperties metadata -->
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-configuration-processor</artifactId>
-			<optional>true</optional>
-		</dependency>
 	</dependencies>
 </project>
