+++ spring-cloud-gcp-starters/spring-cloud-gcp-starter-sql/src/main/resources/META-INF/spring.provides
@@ -1,0 +1,1 @@
+provides: spring-cloud-gcp-starter-core, HikariCP, mysql-socket-factory, mysql-connector-java, postgres-socket-factory, postgresql
