--- spring-cloud-gcp-starters/spring-cloud-gcp-starter-sql/src/test/java/org/springframework/cloud/gcp/sql/autoconfig/GcpCloudSqlTestConfiguration.java
@@ -1,39 +1,0 @@
-/*
- *  Copyright 2017 original author or authors.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *       http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package org.springframework.cloud.gcp.sql.autoconfig;
-
-import com.google.api.gax.core.CredentialsProvider;
-import com.google.auth.Credentials;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-import static org.mockito.Mockito.mock;
-
-/**
- * Provides test mocks.
- *
- * @author João André Martins
- */
-@Configuration
-public class GcpCloudSqlTestConfiguration {
-
-	@Bean
-	public CredentialsProvider googleCredentials() {
-		return () -> mock(Credentials.class);
-	}
-}
