--- spring-cloud-gcp-starters/spring-cloud-gcp-starter-storage/src/main/resources/META-INF/spring.factories
@@ -1,2 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.cloud.gcp.storage.autoconfig.StorageAutoConfiguration
