--- spring-cloud-gcp-starters/spring-cloud-gcp-starter-trace/src/main/resources/META-INF/spring.factories
@@ -1,3 +1,0 @@
-# Auto Configuration
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.springframework.cloud.gcp.trace.autoconfig.StackdriverTraceAutoConfiguration
