+++ spring-cloud-gcp-starters/spring-cloud-gcp-starter-trace/src/main/resources/META-INF/spring.provides
@@ -1,0 +1,1 @@
+provides: spring-cloud-gcp-starter-core, spring-cloud-starter-sleuth, spring-cloud-gcp-trace, gax
